requirements.txt
Django==4.1.7
mysqlclient==2.1.1
python-dotenv==1.0.0
django-haystack==3.2.1
GDAL==3.4.3
Pillow==9.5.0
djangorestframework==3.14.0